+++ libuavcan/include/uavcan/util/lazy_constructor.hpp
@@ -60,13 +60,13 @@ public:
     LazyConstructor()
         : ptr_(NULL)
     {
-        fill(data_.pool, data_.pool + sizeof(T), 0);
+        fill(data_.pool, data_.pool + sizeof(T), uint8_t(0));
     }
 
     LazyConstructor(const LazyConstructor<T>& rhs)   // Implicit
         : ptr_(NULL)
     {
-        fill(data_.pool, data_.pool + sizeof(T), 0);
+        fill(data_.pool, data_.pool + sizeof(T), uint8_t(0));
         if (rhs)
         {
             construct<const T&>(*rhs);  // Invoke copy constructor
@@ -102,7 +102,7 @@ public:
             ptr_->~T();
         }
         ptr_ = NULL;
-        fill(data_.pool, data_.pool + sizeof(T), 0);
+        fill(data_.pool, data_.pool + sizeof(T), uint8_t(0));
     }
 
     void construct()
